Merge pull request #241 from PavelSindler/slow_printer_fix

+++ Firmware/Marlin.h
@@ -96,7 +96,7 @@ void serial_echopair_P(const char *s_P, float v);
 void serial_echopair_P(const char *s_P, double v);
 void serial_echopair_P(const char *s_P, unsigned long v);
 
-extern int selectedSerialPort;
+extern uint8_t selectedSerialPort;
 
 //Things to write to serial from Program memory. Saves 400 to 2k of RAM.
 FORCE_INLINE void serialprintPGM(const char *str)
