+++ Firmware/Marlin_main.cpp
@@ -297,7 +297,7 @@ unsigned int custom_message_type;
 unsigned int custom_message_state;
 char snmm_filaments_used = 0;
 
-int selectedSerialPort;
+uint8_t selectedSerialPort;
 
 float distance_from_min[3];
